PestTracker2/requirements.txt
gradio==3.24.1
torch==2.0.1
ultralytics==8.0.121
opencv-python-headless==4.8.0.74
numpy==1.24.4
Pillow==10.0.0
